Refactor: unify paged_attention golden implementations
Extract common paged_attention golden logic into shared module to reduce
code duplication across examples and tests.

- Add golden/paged_attention_golden.py with shared generate_inputs(),
  paged_attention(), compute_golden(), and run_golden_test() functions
- Refactor 7 golden.py files to import from shared module instead of
  duplicating ~200 lines each
- Each variant now only defines its specific ALL_CASES, RTOL/ATOL, and
  return_all_sizes configuration
- Update examples/scripts/run_example.py to add golden/ to sys.path
- Net reduction: 1643 deletions, 360 additions (-1283 lines)

Affected files:
- golden/paged_attention_golden.py (new shared module)
- examples/scripts/run_example.py
- examples/host_build_graph/paged_attention/golden.py
- examples/tensormap_and_ringbuffer/{paged_attention,batch_paged_attention}/golden.py
- tests/device_tests/aicpu_build_graph/paged_attention/golden.py
- tests/device_tests/host_build_graph/paged_attention/golden.py
- tests/device_tests/tensormap_and_ringbuffer/{paged_attention,batch_paged_attention}/golden.py
